+++ packages/@tinacms/graphql/src/primitives/resolver/index.ts
@@ -134,6 +134,60 @@ export class Resolver {
       throw e
     }
   }
+
+  public getDocumentFields = async () => {
+    try {
+      const response = {}
+      const collections = await this.tinaSchema.getCollections()
+
+      /**
+       * Iterate through collections...
+       */
+      await sequential(collections, async (collection) => {
+        const collectable =
+          this.tinaSchema.getTemplatesForCollectable(collection)
+
+        switch (collectable.type) {
+          /**
+           * Collection with no templates...
+           */
+          case 'object':
+            response[collection.name] = {
+              fields: await sequential(
+                collectable.template.fields,
+                async (field) => {
+                  return this.resolveField(field)
+                }
+              ),
+            }
+            break
+          /**
+           * Collection with n templates...
+           */
+          case 'union':
+            const templates = {}
+            /**
+             * Iterate through templates...
+             */
+            await sequential(collectable.templates, async (template) => {
+              templates[lastItem(template.namespace)] = {
+                fields: await sequential(template.fields, async (field) => {
+                  return this.resolveField(field)
+                }),
+              }
+            })
+
+            response[collection.name] = { templates }
+            break
+        }
+      })
+
+      return response
+    } catch (e) {
+      throw e
+    }
+  }
+
   public resolveDocument = async ({
     value,
     args,
@@ -203,6 +257,12 @@ export class Resolver {
               _template: lastItem(template.namespace),
             })
             return this.getDocument(realPath)
+        }
+      } else {
+        if (!(await this.database.documentExists(realPath))) {
+          throw new Error(
+            `Unable to update document, ${realPath} does not exist`
+          )
         }
       }
       const templateInfo =