+++ packages/@tinacms/toolkit/src/packages/fields/plugins/GroupFieldPlugin.tsx
@@ -21,7 +21,7 @@ import { Field, Form } from '../../forms'
 import styled, { keyframes, css, StyledComponent } from 'styled-components'
 import { FieldsBuilder, useFormPortal, TinaField } from '../../form-builder'
 import { LeftArrowIcon, RightArrowIcon } from '../../icons'
-import { wrapFieldsWithMeta } from './wrapFieldWithMeta'
+import { useCMS } from '../../react-core/use-cms'
 
 export interface GroupFieldDefinititon extends Field {
   component: 'group'
@@ -54,24 +54,6 @@ export const Group = ({ tinaForm, field }: GroupProps) => {
   )
 }
 
-// export const Group = wrapFieldsWithMeta(({ tinaForm, field }: GroupProps) => {
-//   const [isExpanded, setExpanded] = React.useState<boolean>(false)
-//   return (
-//     <>
-//       <Header onClick={() => setExpanded((p) => !p)}>
-//         {field.label || field.name}
-//         <RightArrowIcon />
-//       </Header>
-//       <Panel
-//         isExpanded={isExpanded}
-//         setExpanded={setExpanded}
-//         field={field}
-//         tinaForm={tinaForm}
-//       />
-//     </>
-//   )
-// })
-
 interface PanelProps {
   setExpanded(next: boolean): void
   isExpanded: boolean
@@ -85,6 +67,7 @@ const Panel = function Panel({
   tinaForm,
   field,
 }: PanelProps) {
+  const cms = useCMS()
   const FormPortal = useFormPortal()
   const fields: any[] = React.useMemo(() => {
     return field.fields.map((subField: any) => ({
@@ -100,7 +83,17 @@ const Panel = function Panel({
           isExpanded={isExpanded}
           style={{ zIndex: zIndexShift + 1000 }}
         >
-          <PanelHeader onClick={() => setExpanded(false)}>
+          <PanelHeader
+            onClick={() => {
+              const state = tinaForm.finalForm.getState()
+              if (state.invalid === true) {
+                // @ts-ignore
+                cms.alerts.error('Cannot navigate away from an invalid form.')
+              } else {
+                setExpanded(false)
+              }
+            }}
+          >
             <LeftArrowIcon /> <span>{field.label || field.name}</span>
           </PanelHeader>
           <PanelBody>