+++ packages/tinacms/src/auth/TinaCloudProvider.tsx
@@ -13,7 +13,14 @@ limitations under the License.
 
 import { ModalBuilder } from './AuthModal'
 import React, { useState } from 'react'
-import { TinaCMS, TinaProvider, MediaStore } from '@tinacms/toolkit'
+import {
+  TinaCMS,
+  TinaProvider,
+  MediaStore,
+  //@ts-ignore why can't it find you
+  BranchSwitcherPlugin,
+  Branch
+} from '@tinacms/toolkit'
 
 import { Client, TinaIOConfig } from '../client'
 import { useTinaAuthRedirect } from './useTinaAuthRedirect'
@@ -148,7 +155,46 @@ export const TinaCloudProvider = (
       cms.media.store = new (await props.mediaStore)(cms.api.tina)
     }
   }
+  const handleListBranches = async (): Promise<Branch[]> => {
+    const { owner, repo } = props
+    const branches = await cms.api.tina.listBranches({ owner, repo })
+    
+    return branches.map((branch) => branch.name)
+  }
+  const handleCreateBranch = async (data) => {
+     const newBranch = await cms.api.tina.createBranch(data)
+
+     return newBranch
+  }
+
   setupMedia()
+
+  //@ts-ignore it's not picking up cms.flags
+  const branchingEnabled = cms.flags.get('branch-switcher')
+  
+  React.useEffect(() => {
+    let branchSwitcher
+    if (branchingEnabled) {
+      branchSwitcher = new BranchSwitcherPlugin({
+        cms,
+        owner: props.owner,
+        repo: props.repo,
+        baseBranch: props.branch || 'main',
+        currentBranch: props.branch || 'main',
+        //TODO implement these
+        listBranches: handleListBranches,
+        createBranch: handleCreateBranch,
+        setCurrentBranch: () => console.log(props.branch)
+      })
+      cms.plugins.add(branchSwitcher)
+    }
+    return () => {
+      if (!branchingEnabled) {
+        cms.plugins.remove(branchSwitcher)
+      }
+    }
+  }, [branchingEnabled, props.branch])
+
   if (props.cmsCallback) {
     props.cmsCallback(cms)
   }