+++ packages/tinacms/src/client/index.ts
@@ -22,10 +22,13 @@ import {
 
 import { formify } from './formify'
 import gql from 'graphql-tag'
-export type TinaIOConfig =
-  | { baseUrl?: string; identityApiUrl: undefined; contentApiUrl: undefined }
-  | { baseUrl: undefined; identityApiUrl?: string; contentApiUrl?: string }
-
+//@ts-ignore can't locate BranchChangeEvent
+import { EventBus, BranchChangeEvent, BranchData } from '@tinacms/toolkit'
+export type TinaIOConfig = {
+  frontendUrlOverride?: string // https://app.tina.io
+  identityApiUrlOverride?: string // https://identity.tinajs.io
+  contentApiUrlOverride?: string // https://content.tinajs.io
+}
 interface ServerOptions {
   clientId: string
   branch: string
@@ -36,6 +39,7 @@ interface ServerOptions {
 }
 
 export class Client {
+  frontendUrl: string
   contentApiUrl: string
   identityApiUrl: string
   schema: GraphQLSchema
@@ -44,24 +48,20 @@ export class Client {
   setToken: (_token: TokenObject) => void
   private getToken: () => TokenObject
   private token: string // used with memory storage
-  private baseUrl: string
+  private branch: string
+  private options: ServerOptions
+  events: EventBus // automatically hooked into global event bus when attached via cms.registerApi
 
   constructor({ tokenStorage = 'MEMORY', ...options }: ServerOptions) {
-    const encodedBranch = encodeURIComponent(options.branch)
-    this.baseUrl = options.tinaioConfig?.baseUrl || 'tinajs.io'
-    if (options.tinaioConfig?.contentApiUrl) {
-      this.contentApiUrl =
-        options.customContentApiUrl || options.tinaioConfig.contentApiUrl
-    } else {
-      this.contentApiUrl =
-        options.customContentApiUrl ||
-        `https://content.${this.baseUrl}/content/${options.clientId}/github/${encodedBranch}`
-    }
-    if (options.tinaioConfig?.identityApiUrl) {
-      this.identityApiUrl = options.tinaioConfig.identityApiUrl
-    } else {
-      this.identityApiUrl = `https://identity.${this.baseUrl}`
-    }
+    this.options = options
+    this.setBranch(options.branch)
+    this.events = new EventBus()
+    this.events.subscribe<BranchChangeEvent>(
+      'branch-switcher:change-branch',
+      ({ branchName }) => {
+        this.setBranch(branchName)
+      }
+    )
 
     this.clientId = options.clientId
 
@@ -108,6 +108,21 @@ export class Client {
         this.getToken = options.getTokenFn
         break
     }
+  }
+
+  setBranch(branchName: string) {
+    const encodedBranch = encodeURIComponent(branchName)
+    this.frontendUrl =
+      this.options.tinaioConfig?.frontendUrlOverride || 'https://app.tina.io'
+    this.identityApiUrl =
+      this.options.tinaioConfig?.identityApiUrlOverride ||
+      'https://identity.tinajs.io'
+    const contentApiBase =
+      this.options.tinaioConfig?.contentApiUrlOverride ||
+      `https://content.tinajs.io`
+    this.contentApiUrl =
+      this.options.customContentApiUrl ||
+      `${contentApiBase}/content/${this.options.clientId}/github/${encodedBranch}`
   }
 
   addPendingContent = async (props) => {
@@ -201,7 +216,7 @@ mutation addPendingDocumentMutation(
   }
 
   async authenticate() {
-    const token = await authenticate(this.clientId)
+    const token = await authenticate(this.clientId, this.frontendUrl)
     this.setToken(token)
     return token
   }
@@ -250,7 +265,42 @@ mutation addPendingDocumentMutation(
       return null
     }
   }
+
+  async listBranches ({ owner, repo }: BranchData) {
+    const url = `${this.contentApiUrl}/list_branches?owner=${owner}&repo=${repo}`
+    try {
+      const res = await this.fetchWithToken(url, {
+        method: 'GET'
+      }) 
+
+      return JSON.stringify(res)
+    } catch (e) {
+      console.error("There was an issue fetching the branch list.", e)
+      return null
+    }
+  }
+  async createBranch ({ owner, repo, baseBranch, branchName}: BranchData) {
+    const url = `${this.contentApiUrl}/create_branch`
+  
+    try {
+      const res = await this.fetchWithToken(url, {
+        method: 'POST',
+        body: {
+          owner,
+          repo,
+          baseBranch,
+          branchName
+        } as any
+      })
+
+      return JSON.stringify(res)
+    } catch (error) {
+      console.error('There was an error creating a new branch.', error)
+      return null
+    }
+  }
 }
+
 
 export const DEFAULT_LOCAL_TINA_GQL_SERVER_URL = 'http://localhost:4001/graphql'