+++ examples/tina-tailwind-sidebar-demo/CHANGELOG.md
@@ -1,5 +1,20 @@
 # tina-tailwind-sidebar-demo
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [b59f23295]
+- Updated dependencies [2908f8176]
+- Updated dependencies [08ef183a0]
+- Updated dependencies [c6e2dd69a]
+- Updated dependencies [5df9fe543]
+- Updated dependencies [3f9cad860]
+- Updated dependencies [aa714b1fb]
+  - react-tinacms-editor@0.52.0
+  - react-tinacms-inline@0.53.0
+  - tinacms@0.56.0
+
 ## 0.1.3
 
 ### Patch Changes